Better cert path validation exception message for PartyAndCertificate.verify
