+++ core/src/main/kotlin/net/corda/core/identity/PartyAndCertificate.kt
@@ -2,6 +2,7 @@ package net.corda.core.identity
 
 import net.corda.core.internal.CertRole
 import net.corda.core.internal.uncheckedCast
+import net.corda.core.internal.validate
 import net.corda.core.serialization.CordaSerializable
 import java.security.PublicKey
 import java.security.cert.*
@@ -40,9 +41,7 @@ class PartyAndCertificate(val certPath: CertPath) {
 
     /** Verify the certificate path is valid. */
     fun verify(trustAnchor: TrustAnchor): PKIXCertPathValidatorResult {
-        val parameters = PKIXParameters(setOf(trustAnchor)).apply { isRevocationEnabled = false }
-        val validator = CertPathValidator.getInstance("PKIX")
-        val result = validator.validate(certPath, parameters) as PKIXCertPathValidatorResult
+        val result = certPath.validate(trustAnchor)
         // Apply Corda-specific validity rules to the chain. This only applies to chains with any roles present, so
         // an all-null chain is in theory valid.
         var parentRole: CertRole? = CertRole.extract(result.trustAnchor.trustedCert)
